Reincorporate the glfwHints
Because without the hints GLFW_CONTEXT_VERSION_MAJOR, 3 and GLFW_CONTEXT_VERSION_MINOR, 3. OpenGL will automatically decide base on your driver the lowest or heights vercion of OpenGL in the device. That could bring a problem were OpenGL tries to use a function that is not in the current vercion due to it being to new or old.

Also the hint GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE is telling OpenGL to use the newest function. So, calls like glBegin(), glOpen(), glVertex3f(), etc. Wouldn´t work in this project